+++ src/test/converter2/issues/gh3019.ts
@@ -1,0 +1,4 @@
+class GH3019 {
+    accessor x: string = "";
+    accessor y: number = 123;
+}